requirements.txt
beautifulsoup4==4.11.1
dnspython==2.2.1
Flask==2.2.2
googletrans==3.0.0
opencc==1.1.4
pymongo==4.2.0
pytube==12.1.0
pytz==2021.3
requests==2.26.0
telegram==0.0.1
wikipedia==1.4.0
